requirements.txt
Flask
pandas
numpy
tensorflow
Pillow